refactor: 데스크톱/모바일 분기 훅 생성
