+++ src/app/hooks/useIsDesktop.tsx
@@ -1,0 +1,21 @@
+"use client";
+
+import { useEffect, useState } from "react";
+
+export function useIsDesktop() {
+  const [isDesktop, setIsDesktop] = useState(false);
+
+  useEffect(() => {
+    if (typeof window === "undefined") return;
+
+    const mq = window.matchMedia("(min-width: 768px)");
+
+    const update = () => setIsDesktop(mq.matches);
+    update();
+
+    mq.addEventListener("change", update);
+    return () => mq.removeEventListener("change", update);
+  }, []);
+
+  return isDesktop;
+}
